# contracts/v1/openapi.json
{
  "components": {
    "schemas": {
      "AdversarialDialectic": {
        "description": "Dual-Agent Dialectic: Adversarial debate synthesizing prosecution and defense arguments.",
        "properties": {
          "defenseMitigation": {
            "description": "Forensic defense advocate highlighting natural compression/sensor noise",
            "title": "Defensemitigation",
            "type": "string"
          },
          "judicialSynthesis": {
            "description": "Balanced judicial ruling justifying the calibrated Trust Vector",
            "title": "Judicialsynthesis",
            "type": "string"
          },
          "prosecutionArgument": {
            "description": "Forensic prosecutor argument detailing anomalies and red flags",
            "title": "Prosecutionargument",
            "type": "string"
          }
        },
        "required": [
          "prosecutionArgument",
          "defenseMitigation",
          "judicialSynthesis"
        ],
        "title": "AdversarialDialectic",
        "type": "object"
      },
...
      "AssessmentTier": {
        "enum": [
          "BENIGN_AUTHENTIC",
          "LOW_RISK_VERIFIED",
          "UNCERTAIN_COMPRESSION_NOISE",
          "SUSPICIOUS_ANOMALY",
          "HIGH_IMPERSONATION_RISK"
        ],
        "title": "AssessmentTier",
        "type": "string"
      },
      "AuditCertificateResponse": {
        "properties": {
          "auditCertificateId": {
            "title": "Auditcertificateid",
            "type": "string"
          },
          "certificateSha256": {
            "title": "Certificatesha256",
            "type": "string"
          },
          "downloadPdfUrl": {
            "title": "Downloadpdfurl",
            "type": "string"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "string"
          },
          "verifyUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verifyurl"
          }
        },
        "required": [
          "auditCertificateId",
          "timestamp",
          "certificateSha256",
          "downloadPdfUrl"
        ],
        "title": "AuditCertificateResponse",
        "type": "object"
      },
      "AuditSignRequest": {
        "properties": {
          "analystId": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Analystid",
            "type": "string"
          },
          "analystNotes": {
            "maxLength": 3000,
            "minLength": 10,
            "title": "Analystnotes",
            "type": "string"
          },
          "completedSteps": {
            "items": {
              "type": "integer"
            },
            "title": "Completedsteps",
            "type": "array"
          },
          "finalDecision": {
            "title": "Finaldecision",
            "type": "string"
          }
        },
        "required": [
          "analystId",
          "analystNotes",
          "finalDecision",
          "completedSteps"
        ],
        "title": "AuditSignRequest",
        "type": "object"
      },
...
      "CanaryTripwireAlert": {
        "description": "Active deception detection via honeypot tokens and zero-width markers.",
        "properties": {
          "details": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "tripwireTriggered": {
            "default": false,
            "title": "Tripwiretriggered",
            "type": "boolean"
          },
          "tripwireType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tripwiretype"
          }
        },
        "title": "CanaryTripwireAlert",
        "type": "object"
      },
      "CertificateVerificationResponse": {
        "properties": {
          "algorithm": {
            "title": "Algorithm",
            "type": "string"
          },
          "certificateSha256": {
            "title": "Certificatesha256",
            "type": "string"
          },
          "manifestSignature": {
            "title": "Manifestsignature",
            "type": "string"
          },
          "notice": {
            "title": "Notice",
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/SignedAuditPayload"
          },
          "payloadSha256": {
            "title": "Payloadsha256",
            "type": "string"
          },
          "publicKey": {
            "title": "Publickey",
            "type": "string"
          },
          "signature": {
            "title": "Signature",
            "type": "string"
          },
          "valid": {
            "title": "Valid",
            "type": "boolean"
          }
        },
        "required": [
          "valid",
          "algorithm",
          "publicKey",
          "payloadSha256",
          "certificateSha256",
          "signature",
          "payload",
          "manifestSignature",
          "notice"
        ],
        "title": "CertificateVerificationResponse",
        "type": "object"
      },
      "ClaimedIdentity": {
        "properties": {
          "canonicalUserId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonicaluserid"
          },
          "claimedAffiliation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claimedaffiliation"
          },
          "claimedRole": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claimedrole"
          },
          "entityName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entityname"
          },
          "referenceHandles": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReferenceHandles"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "ClaimedIdentity",
        "type": "object"
      },
      "EnvironmentalForensics": {
        "description": "Acoustic and physiological environmental micro-signals.",
        "properties": {
          "ambientAcousticProfile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "e.g. Studio Dead, Open Office, Reverb Hallway",
            "title": "Ambientacousticprofile"
          },
          "measurementNotes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Measurementnotes"
          },
          "pulseFrequencyHz": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pulsefrequencyhz"
          },
          "reverberationMatchScore": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "RT60 acoustic space consistency",
            "title": "Reverberationmatchscore"
          },
          "rppgPulseConfidence": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Legacy field name: chrominance periodicity, NOT biological liveness probability",
            "title": "Rppgpulseconfidence"
          },
          "rt60Sec": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rt60Sec"
          }
        },
        "title": "EnvironmentalForensics",
        "type": "object"
      },
      "EvidenceItem": {
        "properties": {
          "evidenceId": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Evidenceid",
            "type": "string"
          },
          "mediaUri": {
            "anyOf": [
              {
                "maxLength": 2048,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mediauri"
          },
          "metadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvidenceMetadata"
              },
              {
                "type": "null"
              }
            ]
          },
          "modality": {
            "$ref": "#/components/schemas/ModalityType"
          },
          "samples": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvidenceSamples"
              },
              {
                "type": "null"
              }
            ]
          },
          "textContent": {
            "anyOf": [
              {
                "maxLength": 20000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Textcontent"
          }
        },
        "required": [
          "evidenceId",
          "modality"
        ],
        "title": "EvidenceItem",
        "type": "object"
      },
...
      "HTTPErrorResponse": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "$ref": "#/components/schemas/ValidationIssue"
                },
                "type": "array"
              }
            ],
            "title": "Detail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "HTTPErrorResponse",
        "type": "object"
      },
...
      "LedgerEvidenceItem": {
        "properties": {
          "affectedSpan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AffectedSpan"
              },
              {
                "type": "null"
              }
            ]
          },
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "finding": {
            "title": "Finding",
            "type": "string"
          },
          "layer": {
            "$ref": "#/components/schemas/ForensicLayer"
          },
          "modality": {
            "title": "Modality",
            "type": "string"
          },
          "polarity": {
            "$ref": "#/components/schemas/EvidencePolarity"
          },
          "signalId": {
            "title": "Signalid",
            "type": "string"
          }
        },
        "required": [
          "signalId",
          "layer",
          "modality",
          "polarity",
          "confidence",
          "finding"
        ],
        "title": "LedgerEvidenceItem",
        "type": "object"
      },
...
      "ScenarioDetailResponse": {
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "request": {
            "$ref": "#/components/schemas/TrustGuardInspectionRequest"
          },
          "scenarioId": {
            "title": "Scenarioid",
            "type": "string"
          },
          "simulation": {
            "default": true,
            "title": "Simulation",
            "type": "boolean"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "scenarioId",
          "title",
          "description",
          "request"
        ],
        "title": "ScenarioDetailResponse",
        "type": "object"
      },
      "ScenarioSummary": {
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "scenarioId": {
            "title": "Scenarioid",
            "type": "string"
          },
          "simulation": {
            "default": true,
            "title": "Simulation",
            "type": "boolean"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "scenarioId",
          "title",
          "description"
        ],
        "title": "ScenarioSummary",
        "type": "object"
      },
      "SignedAuditPayload": {
        "properties": {
          "analystStatement": {
            "$ref": "#/components/schemas/AuditSignRequest"
          },
          "certificateId": {
            "title": "Certificateid",
            "type": "string"
          },
          "scope": {
            "title": "Scope",
            "type": "string"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "string"
          },
          "verdict": {
            "$ref": "#/components/schemas/TrustGuardInspectionVerdict"
          }
        },
        "required": [
          "certificateId",
          "timestamp",
          "verdict",
          "analystStatement",
          "scope"
        ],
        "title": "SignedAuditPayload",
        "type": "object"
      },
      "SourceChannel": {
        "enum": [
          "web_portal",
          "browser_extension",
          "headless_api",
          "citizen_tipline"
        ],
        "title": "SourceChannel",
        "type": "string"
      },
      "TrustGuardInspectionRequest": {
        "properties": {
          "claimedIdentity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClaimedIdentity"
              },
              {
                "type": "null"
              }
            ]
          },
          "evidenceItems": {
            "items": {
              "$ref": "#/components/schemas/EvidenceItem"
            },
            "maxItems": 8,
            "minItems": 1,
            "title": "Evidenceitems",
            "type": "array"
          },
          "requestId": {
            "description": "Unique inspection request identifier",
            "maxLength": 128,
            "minLength": 1,
            "title": "Requestid",
            "type": "string"
          },
          "sourceChannel": {
            "$ref": "#/components/schemas/SourceChannel"
          },
          "timestamp": {
            "description": "ISO 8601 timestamp of inspection request",
            "maxLength": 64,
            "minLength": 1,
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "requestId",
          "timestamp",
          "sourceChannel",
          "evidenceItems"
        ],
        "title": "TrustGuardInspectionRequest",
        "type": "object"
      },
      "TrustGuardInspectionVerdict": {
        "properties": {
          "adversarialDialectic": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AdversarialDialectic"
              },
              {
                "type": "null"
              }
            ]
          },
          "assessmentTier": {
            "$ref": "#/components/schemas/AssessmentTier"
          },
          "calibratedTrustVector": {
            "$ref": "#/components/schemas/TrustVector"
          },
          "environmentalForensics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EnvironmentalForensics"
              },
              {
                "type": "null"
              }
            ]
          },
          "evaluatedAt": {
            "title": "Evaluatedat",
            "type": "string"
          },
          "evidenceLedger": {
            "items": {
              "$ref": "#/components/schemas/LedgerEvidenceItem"
            },
            "title": "Evidenceledger",
            "type": "array"
          },
          "innovationMetadata": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Innovationmetadata"
          },
          "requestId": {
            "title": "Requestid",
            "type": "string"
          },
          "tripwireStatus": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CanaryTripwireAlert"
              },
              {
                "type": "null"
              }
            ]
          },
          "verdictId": {
            "title": "Verdictid",
            "type": "string"
          },
          "verdictSummary": {
            "$ref": "#/components/schemas/VerdictSummary"
          },
          "verificationPlaybook": {
            "items": {
              "$ref": "#/components/schemas/PlaybookStep"
            },
            "title": "Verificationplaybook",
            "type": "array"
          }
        },
        "required": [
          "verdictId",
          "requestId",
          "evaluatedAt",
          "assessmentTier",
          "calibratedTrustVector",
          "verdictSummary",
          "evidenceLedger",
          "verificationPlaybook"
        ],
        "title": "TrustGuardInspectionVerdict",
        "type": "object"
      },
      "TrustVector": {
        "properties": {
          "contextualAnomalyScore": {
            "description": "Contextual/historical anomaly score [0,1]",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Contextualanomalyscore",
            "type": "number"
          },
          "crossModalDiscordanceScore": {
            "description": "Cross-modal discordance index [0,1]",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Crossmodaldiscordancescore",
            "type": "number"
          },
          "epistemicUncertainty": {
            "description": "Quality degradation/uncertainty metric [0,1]",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Epistemicuncertainty",
            "type": "number"
          },
          "identityMismatchScore": {
            "description": "Impersonation drift vs baseline [0,1]",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Identitymismatchscore",
            "type": "number"
          },
          "mediaSynthesisScore": {
            "description": "Physical/spectral synthesis score [0,1]",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Mediasynthesisscore",
            "type": "number"
          }
        },
        "required": [
          "mediaSynthesisScore",
          "crossModalDiscordanceScore",
          "identityMismatchScore",
          "contextualAnomalyScore",
          "epistemicUncertainty"
        ],
        "title": "TrustVector",
        "type": "object"
      },
      "ValidationIssue": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Loc",
            "type": "array"
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationIssue",
        "type": "object"
      },
      "VerdictSummary": {
        "properties": {
          "coreAnomaly": {
            "title": "Coreanomaly",
            "type": "string"
          },
          "headline": {
            "title": "Headline",
            "type": "string"
          },
          "noScoreIsProofNotice": {
            "default": "Under TrustGuard core architecture, no single score constitutes legal or definitive proof. Review the calibrated 5D Trust Vector and complete the Human Verification Playbook before taking adverse action.",
            "title": "Noscoreisproofnotice",
            "type": "string"
          }
        },
        "required": [
          "headline",
          "coreAnomaly"
        ],
        "title": "VerdictSummary",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Multi-Modal AI for Digital Trust · Track 01 PS-02 Core Engine",
    "title": "TrustGuard",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/cases/{case_id}/sign-audit": {
      "post": {
        "operationId": "sign_audit_api_v1_cases__case_id__sign_audit_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuditSignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditCertificateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Requested scenario, case or certificate was not found."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "The request exceeds the transient body-size limit."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Validation failed; detail is a message or a list of sanitized validation issues."
          }
        },
        "summary": "Sign Audit",
        "tags": [
          "Human Verification"
        ]
      }
    },
    "/api/v1/certificates/{certificate_id}.pdf": {
      "get": {
        "operationId": "download_certificate_api_v1_certificates__certificate_id__pdf_get",
        "parameters": [
          {
            "in": "path",
            "name": "certificate_id",
            "required": true,
            "schema": {
              "title": "Certificate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/pdf": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Signed incident certificate PDF"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Requested scenario, case or certificate was not found."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "The request exceeds the transient body-size limit."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Validation failed; detail is a message or a list of sanitized validation issues."
          }
        },
        "summary": "Download Certificate",
        "tags": [
          "Human Verification"
        ]
      }
    },
    "/api/v1/certificates/{certificate_id}/verify": {
      "get": {
        "operationId": "verify_certificate_api_v1_certificates__certificate_id__verify_get",
        "parameters": [
          {
            "in": "path",
            "name": "certificate_id",
            "required": true,
            "schema": {
              "title": "Certificate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CertificateVerificationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Requested scenario, case or certificate was not found."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "The request exceeds the transient body-size limit."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Validation failed; detail is a message or a list of sanitized validation issues."
          }
        },
        "summary": "Verify Certificate",
        "tags": [
          "Human Verification"
        ]
      }
    },
...
    "/api/v1/scenarios": {
      "get": {
        "operationId": "list_scenarios_api_v1_scenarios_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScenarioSummary"
                  },
                  "title": "Response List Scenarios Api V1 Scenarios Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Requested scenario, case or certificate was not found."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "The request exceeds the transient body-size limit."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Validation failed; detail is a message or a list of sanitized validation issues."
          }
        },
        "summary": "List Scenarios",
        "tags": [
          "Scenarios"
        ]
      }
    },
    "/api/v1/scenarios/{scenario_id}": {
      "get": {
        "operationId": "get_scenario_api_v1_scenarios__scenario_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Requested scenario, case or certificate was not found."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "The request exceeds the transient body-size limit."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Validation failed; detail is a message or a list of sanitized validation issues."
          }
        },
        "summary": "Get Scenario",
        "tags": [
          "Scenarios"
        ]
      }
    }
  }
}